+++ examples/demo/demo.cpp
@@ -21,7 +21,9 @@
 
 #include <Resource.h>
 
+#ifdef _WIN32
 #include <tchar.h>
+#endif
 
 #ifdef _WIN32
 #include <gdiplus.h>
@@ -34,7 +36,7 @@ int APIENTRY _tWinMain(_In_ HINSTANCE,
     _In_opt_ HINSTANCE,
     _In_ LPTSTR    lpCmdLine,
     _In_ int       nCmdShow)
-#elif __linux__
+#elif defined(__linux__) || defined(__APPLE__)
 int main(int argc, char *argv[])
 #endif
 {